pyproject.toml
[tool.poetry]
name = "Hackathon-backend"
version = "0.1.0"
description = ""
authors = ["William Lindholm <william.lindholm@ander.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.0.5"
djangorestframework = "^3.13.1"
gunicorn = "^20.1.0"
psycopg2 = "^2.9.3"
whitenoise = "^6.2.0"
drf-yasg = "^1.20.0"
sentry-sdk = "^1.5.12"
django-environ = "^0.8.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
exclude = "(migrations|node_modules)"
line_length = 120

[tool.isort]
profile = "black"
line_length=120
atomic = true
skip_glob = ["*/migrations/*", "*/node_modules/*"]
